Rename files to match module names
DUB builds each package separately, and then adds each package to the
import paths using D's usual mechanism of generating the filesystem
path name from the module name.

To make this work, source files must be in a path that matches the
module name. This commit makes no changes to the files, it just moves
them to the correct path for discovery by import.