Resolve name conflict from GDC libraries
GDC version of Phobos defines a deprecated version of indexOf in
std.algorithm.  The suggested fix is to use std.algorithm.countUntil,
but as we're dealing with pure ASCII text, the original std.string.indexOf
is fine for our purposes. So we simply name usage of indexOf with a fully
qualified name.